requirements.txt
runpod
pillow
fastapi
uvicorn
numpy
transformers
diffusers
accelerate
safetensors
opencv-python-headless
